fix(cmd/influxd): prevent panic on upgrade with V1 users and no V1 conf (#20558)
